base - if group is static don't delete

+++ cloudsystem-base/src/main/java/de/bytemc/cloud/services/LocalService.java
@@ -175,6 +175,7 @@ public class LocalService implements IService {
     }
 
     public void delete() {
+        if (this.group.isStatic()) return;
         try {
             FileUtils.deleteDirectory(this.workingDirectory);
         } catch (IOException e) {
